select .db based on uid
Since some of the other settings like Calendar, Contacts, etc. are
stored in the user's library folder, I am trying to set the path based
on that.  Because of that, i remove the sudo_required() function.  If
you are root or using sudo, it will be assumed you want to modify the
accessibility database.  If you are a regular user, the other database
will be modified.